# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurant-portal/restaurants/{restaurantId}/packages:
    get:
      operationId: getRestaurantPackages
      tags: [restaurant-portal-packages]
      summary: Get all meal packages for a restaurant
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RestaurantIdParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [all, active, paused, archived]
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RestaurantPackage"
    post:
      operationId: createRestaurantPackage
      tags: [restaurant-portal-packages]
      summary: Create a new subscription package
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RestaurantIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackageBody"
      responses:
        "201":
          description: Package created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantPackage"
  /restaurant-portal/restaurants/{restaurantId}/packages/{packageId}:
    put:
      operationId: updateRestaurantPackage
      tags: [restaurant-portal-packages]
      summary: Update an existing subscription package
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RestaurantIdParam"
        - name: packageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackageBody"
      responses:
        "200":
          description: Package updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantPackage"
    delete:
      operationId: deleteRestaurantPackage
      tags: [restaurant-portal-packages]
      summary: Archive/delete a subscription package
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RestaurantIdParam"
        - name: packageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Package archived
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
  /restaurant-portal/restaurants/{restaurantId}/meals:
    get:
      operationId: getRestaurantMeals
      tags: [restaurant-portal-meals]
      summary: Get all meals/products for a restaurant
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RestaurantIdParam"
        - name: isActive
          in: query
          schema:
            type: string
            enum: [all, true, false]
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of meals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RestaurantMeal"
    post:
      operationId: createRestaurantMeal
      tags: [restaurant-portal-meals]
      summary: Create a new meal/product
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RestaurantIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMealBody"
      responses:
        "201":
          description: Meal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantMeal"
  /restaurant-portal/restaurants/{restaurantId}/meals/{mealId}:
    put:
      operationId: updateRestaurantMeal
      tags: [restaurant-portal-meals]
      summary: Update an existing meal/product
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RestaurantIdParam"
        - name: mealId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMealBody"
      responses:
        "200":
          description: Meal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantMeal"
    delete:
      operationId: deleteRestaurantMeal
      tags: [restaurant-portal-meals]
      summary: Delete a meal/product
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RestaurantIdParam"
        - name: mealId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Meal deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
...
  /student/restaurants:
    get:
      operationId: listStudentRestaurants
      tags: [student-restaurants]
      summary: List active restaurants
      security:
        - bearerAuth: []
      parameters:
        - name: campusId
          in: query
          schema:
            type: string
        - name: slot
          in: query
          schema:
            type: string
            enum: [lunch, dinner, both]
      responses:
        "200":
          description: Restaurant list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentRestaurant"
  /student/restaurants/{restaurantId}:
    get:
      operationId: getStudentRestaurant
      tags: [student-restaurants]
      summary: Get restaurant detail with packages and meals
      security:
        - bearerAuth: []
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Restaurant detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentRestaurantDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RestaurantIdParam:
      name: restaurantId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    RestaurantPackage:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        mealSlot:
          type: string
          enum: [lunch, dinner, both]
        mealCount:
          type: integer
        pricePerDay:
          type: number
        totalPrice:
          type: number
        validityDays:
          type: integer
        activeSubscribers:
          type: integer
        totalSold:
          type: integer
        revenueGenerated:
          type: number
        status:
          type: string
          enum: [active, paused, archived]
        discountPct:
          type: number
        description:
          type: string
      required:
        - id
        - name
        - mealSlot
        - mealCount
        - pricePerDay
        - totalPrice
        - validityDays
        - activeSubscribers
        - totalSold
        - revenueGenerated
        - status
    CreatePackageBody:
      type: object
      properties:
        name:
          type: string
        mealSlot:
          type: string
          enum: [lunch, dinner, both]
        validityDays:
          type: integer
          minimum: 1
        pricePerDay:
          type: number
          minimum: 0
        discountPct:
          type: number
          minimum: 0
          maximum: 100
        description:
          type: string
        status:
          type: string
          enum: [active, paused, archived]
      required:
        - name
        - mealSlot
        - validityDays
        - pricePerDay
    RestaurantMeal:
      type: object
      properties:
        id:
          type: string
        restaurantId:
          type: string
        name:
          type: string
        shortDescription:
          type: string
        description:
          type: string
        category:
          type: string
        vegType:
          type: string
          enum: [veg, non-veg, egg]
        calories:
          type: integer
        protein:
          type: string
        carbs:
          type: string
        fat:
          type: string
        fiber:
          type: string
        allergens:
          type: string
        spiceLevel:
          type: string
          enum: [mild, medium, hot]
        imageUrl:
          type: string
        isAvailableForLunch:
          type: boolean
        isAvailableForDinner:
          type: boolean
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - restaurantId
        - name
        - vegType
        - isAvailableForLunch
        - isAvailableForDinner
        - isActive
        - sortOrder
        - createdAt
        - updatedAt
    CreateMealBody:
      type: object
      properties:
        name:
          type: string
        shortDescription:
          type: string
        description:
          type: string
        category:
          type: string
        vegType:
          type: string
          enum: [veg, non-veg, egg]
        calories:
          type: integer
        protein:
          type: string
        carbs:
          type: string
        fat:
          type: string
        fiber:
          type: string
        allergens:
          type: string
        spiceLevel:
          type: string
          enum: [mild, medium, hot]
        imageUrl:
          type: string
        isAvailableForLunch:
          type: boolean
        isAvailableForDinner:
          type: boolean
        isActive:
          type: boolean
        sortOrder:
          type: integer
      required:
        - name
...
    StudentRestaurant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        tagline:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        cuisineType:
          type: string
          nullable: true
        campusId:
          type: string
          nullable: true
        isVeg:
          type: boolean
        lunchAvailable:
          type: boolean
        dinnerAvailable:
          type: boolean
        operatingDays:
          type: array
          items:
            type: string
        maxCapacity:
          type: integer
          nullable: true
        accentColor:
          type: string
          nullable: true
        lunchStartPrice:
          type: number
          nullable: true
        dinnerStartPrice:
          type: number
          nullable: true
        cancelCutoffHour:
          type: integer
        deliveryTime:
          type: string
          nullable: true
        distanceLabel:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        reviewCount:
          type: integer
      required: [id, name, isVeg, lunchAvailable, dinnerAvailable, operatingDays, cancelCutoffHour, reviewCount]
    StudentPackage:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        durationDays:
          type: integer
        mealSlot:
          type: string
          enum: [lunch, dinner, both]
        pricePerDay:
          type: number
        totalPrice:
          type: number
        discountPercent:
          type: number
        mealsPerDay:
          type: integer
      required: [id, name, durationDays, mealSlot, pricePerDay, totalPrice, discountPercent, mealsPerDay]
    StudentMeal:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        shortDescription:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        vegType:
          type: string
          enum: [veg, non-veg, egg]
        spiceLevel:
          type: string
          enum: [mild, medium, hot]
          nullable: true
        calories:
          type: integer
          nullable: true
        storagePath:
          type: string
          nullable: true
        lunchAvailable:
          type: boolean
        dinnerAvailable:
          type: boolean
      required: [id, name, vegType, lunchAvailable, dinnerAvailable]
    StudentRestaurantDetail:
      allOf:
        - $ref: "#/components/schemas/StudentRestaurant"
        - type: object
          properties:
            packages:
              type: array
              items:
                $ref: "#/components/schemas/StudentPackage"
            meals:
              type: array
              items:
                $ref: "#/components/schemas/StudentMeal"
            ratingsBreakdown:
              type: object
              nullable: true
              properties:
                totalReviews:
                  type: integer
                overall:
                  type: number
                  nullable: true
                foodQuality:
                  type: number
                  nullable: true
                packaging:
                  type: number
                  nullable: true
                delivery:
                  type: number
                  nullable: true
                valueForMoney:
                  type: number
                  nullable: true
                hygiene:
                  type: number
                  nullable: true
          required: [packages, meals]
...